Added WhileExpression, EscapeException and break and return functions
WhileExpression has the same syntax as IfExpression and the middle operand
is executed untill the first one returns false.  The result of the
expression is the last operand.  During execution of a loop two magic
constants are available: it and last.

EscapeException is a base for BreakException and ReturnException which allow
breaking out from loops and functions.  The break and return functions
simply throw one of those exceptions.

Also a doc/grammar.txt was added with grammar in EBNF format.

There was also some minor clean up after covnerting repository from CVS.
Most notably the "$Id$" lines were removed.